+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { Dialog, DialogTitle } from "@mui/material";
+import { Dialog, DialogTitle, ThemeProvider, createTheme } from "@mui/material";
 import * as React from "react";
 import { useCoreDataLake } from "~CoreDataLake";
 import NavBar from "~NavBar";
@@ -22,19 +22,28 @@ export default () => {
     return () => window.removeEventListener("keydown", keydownListener);
   }, [doSave])
 
-
+  const theme = createTheme({
+    typography: {
+      fontFamily: [
+        'Noto Sans', 'Arial', 'Helvetica', 'sans-serif'
+      ].join(","),
+      fontSize: 12,
+    },
+  })
   return <div style={{ display: "flex", flexDirection: "column", height: "100%" }}>
-    <Dialog open={!dataAndLoadState.loaded} >
-      <DialogTitle>Loading your document...</DialogTitle>
-    </Dialog>
-    <NavBar
-      kvStores={kvStores}
-      setKVStores={setKVStores}
-      dataAndLoadState={dataAndLoadState}
-      setData={updateData}
-    ></NavBar>
-    <div className="viewContainer">
-      <Workflowish data={dataAndLoadState.data} updateData={updateData}></Workflowish>
-    </div>
+    <ThemeProvider theme={theme}>
+      <Dialog open={!dataAndLoadState.loaded} >
+        <DialogTitle>Loading your document...</DialogTitle>
+      </Dialog>
+      <NavBar
+        kvStores={kvStores}
+        setKVStores={setKVStores}
+        dataAndLoadState={dataAndLoadState}
+        setData={updateData}
+      ></NavBar>
+      <div className="viewContainer">
+        <Workflowish data={dataAndLoadState.data} updateData={updateData}></Workflowish>
+      </div>
+    </ThemeProvider>
   </div>
 };