+++ docs/changelog.md
@@ -1,6 +1,7 @@
 # Latest 
 - Add workflowish item creation in script engine
 - Refactor the help screen so it uses workflowky + add symlink info to help screen
+- Fix rearrangement with symlinks
 
 # Version 3.2.0: Rudimentary symlinks 
 - Rudimentary (readonly) symlinks