+++ docs/changelog.md
@@ -2,6 +2,7 @@
 - Add workflowish item creation in script engine
 - Refactor the help screen so it uses workflowky + add symlink info to help screen
 - Fix rearrangement with symlinks
+- Allow created items to have custom ID
 
 # Version 3.2.0: Rudimentary symlinks 
 - Rudimentary (readonly) symlinks