+++ docs/changelog.md
@@ -1,4 +1,4 @@
-# Latest 
+# Version 3.3.0: Miscellaneous Bugfixes 
 - Add workflowish item creation in script engine
 - Refactor the help screen so it uses workflowky + add symlink info to help screen
 - Fix rearrangement with symlinks
@@ -6,7 +6,8 @@
 - Fix pre-commit hook not failing
 - Fix console warning about nested elements
 - Symlinked Nodes should not be readonly, only children should be readonly
-- check for duplication in rendering process to reduce stray items
+- Check for duplication in rendering process to reduce stray items
+- Password prefix for better security
 
 # Version 3.2.0: Rudimentary symlinks 
 - Rudimentary (readonly) symlinks